refactor(core): [P] modernize SQLAlchemy mappings in btypes, connection
(partial: 6112fa77, 74e1d9f7, 635a7d8f, 54ba0dd8)

Replay search branch commit 6f4d2b78b981 on the clean 3.1 development base with a standardized message. The rewritten subject summarizes this as: modernize SQLAlchemy mappings in btypes, connection manager, and database layer. It is part of the SQLAlchemy modernization work.

Original subject: Checkpoint
Original author: Chris Wyse <chris.wyse@wysechoice.net>
Commit scope: core
Changed paths: bauble/__init__.py, bauble/btypes.py, bauble/connmgr.py, bauble/db.py, bauble/editor.py, bauble/meta.py, bauble/pictures_view.py, bauble/pluginmgr.py, bauble/plugins/abcd/__init__.py, bauble/plugins/abcd/test.py, bauble/plugins/garden/__init__.py, bauble/plugins/garden/accession.py, and 55 more

Original search commit: 6f4d2b78b981ab24b338166bd7174bc04bbfdf74

+++ bauble/pluginmgr.py
@@ -56,7 +56,7 @@ from sqlalchemy import Unicode
 
 
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
+logger.setLevel(logging.WARNING)
 
 
 gi.require_version("Gtk", "3.0")
@@ -405,9 +405,12 @@ class PluginRegistry(db.Base):
         if name is None:
             name = plugin.__class__.__name__
         with db.Session() as session:
-            p = session.execute(select(PluginRegistry)).scalars().where(name=utils.utf8(name)).one()
-            session.delete(p)
-            session.commit()
+            p = session.execute(
+                select(PluginRegistry).where(PluginRegistry.name==utils.utf8(name))
+            ).scalar_one_or_none()
+            if p:
+                session.delete(p)
+                session.commit()
 
     @staticmethod
     def all(session):
